requirements.txt
opencv-python>=4.1.0.25
dnfal>=0.8.4
numpy>=1.16.0
gpustat>=0.6.0
requests>=2.22.0
cvtlib>=0.8.5
psutil>=5.6.3
Django>=3.0.2
PyJWT>=1.7.1
djangorestframework==3.10.2
celery>=4.3.0
torch>=1.3.0
drf-yasg>=1.17.0
django-cors-headers>=3.2.1
Pillow>=6.1.0
psycopg2>=2.8.4
pynvml>=8.0.3
fabric>=2.5.0

